# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "L5 Swagger API",
        "description": "L5 Swagger API description",
        "contact": {
            "email": "freddiemokwena949@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
        "/api/item": {
            "get": {
                "summary": "Get all items",
                "operationId": "item",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "put": {
                "summary": "Edit a item",
                "operationId": "item",
                "parameters": [
                    {
                        "name": "I_Name",
                        "in": "path",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "I_Weight",
                        "in": "path",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "I_Price",
                        "in": "path",
                        "required": false,
                        "type": "number"
                    },
                    {
                        "name": "supermarket_id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "post": {
                "summary": "Post a item",
                "operationId": "item",
                "parameters": [
                    {
                        "name": "I_Name",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "I_Weight",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "I_Price",
                        "in": "path",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "supermarket_id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/item/{id}": {
            "get": {
                "summary": "Get a item",
                "operationId": "item",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "delete": {
                "summary": "Delete a item",
                "operationId": "supermarket",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "summary": "Login",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "summary": "Register",
                "operationId": "Register",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "surname",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "address",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "phonenumber",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
        "/api/supermarket": {
            "get": {
                "summary": "Get all supermarkets",
                "operationId": "supermarket",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "post": {
                "summary": "Post a supermarket",
                "operationId": "supermarket",
                "parameters": [
                    {
                        "name": "S_Name",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "_Slocation",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "S_Status",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/supermarket/{id}": {
            "get": {
                "summary": "Get a supermarket",
                "operationId": "supermarket",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "put": {
                "summary": "Edit a supermarket",
                "operationId": "supermarket",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "S_Name",
                        "in": "path",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "_Slocation",
                        "in": "path",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "S_Status",
                        "in": "path",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "delete": {
                "summary": "Delete a supermarket",
                "operationId": "supermarket",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        }
    },
...
}